ai-engine/requirements.txt
# AI Engine Dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
httpx==0.26.0
opencv-python-headless==4.9.0.80
numpy==1.26.3
torch==2.1.2
torchvision==0.16.2
ultralytics==8.1.0
pyaudio==0.2.14
websockets==12.0
Pillow==10.2.0
